# generated/go/model_buzz.go
/*
 * Scaffold Servise
 *
 * Base for other projects
 *
 * API version: 0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Buzz struct {

	Name string `json:"name,omitempty"`

	BuzzType string `json:"buzzType,omitempty"`

	Description string `json:"description,omitempty"`
}