# generated/go/model_couch.go
/*
 * Scaffold Servise
 *
 * Base for other projects
 *
 * API version: 0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Couch struct {

	Id int64 `json:"id,omitempty"`

	FirstName string `json:"firstName,omitempty"`

	SecondName string `json:"secondName,omitempty"`
}