# generated/go/model_health.go
/*
 * Scaffold Servise
 *
 * Base for other projects
 *
 * API version: 0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Health struct {

	// app name
	App string `json:"app,omitempty"`

	// app version
	Version string `json:"version,omitempty"`

	// app tag
	Tag string `json:"tag,omitempty"`

	// app build
	Build string `json:"build,omitempty"`
}